+++ changelog.md
@@ -8,6 +8,7 @@
 - Added some temporary patches to fix a handful of compatibility issues with JuneLib until it gets properly updated
 - Increased the size of Stereoscope's visual indicators and added extra wedge showing current aim direction for increased clarity
 - Fixed null derefs caused by certain guns trying to call update methods in the Breach when given via Expand the Gungeon's gunball machine
+- Fixed [REDACTED] not having outlines before initial interaction
 
 ## 1.28.1 (2025-10-01)